requirements-dev.txt
aws-cdk-lib==2.204.0
constructs~=10.4.2
